homelab/pyproject.toml
[tool.poetry]
name = "control"
version = "0.1.0"
description = "Controls the server."
authors = ["Alex Povel <git@alexpovel.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
docker = "^4.4.3"
PyYAML = "^5.4.1"
Jinja2 = "^2.11.3"
pexpect = "^4.8.0"
docker-compose = "1.27.4"
"backports.zoneinfo" = "^0.2.1"
tzdata = "^2021.1"
ipinfo = "^4.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
